refactor: change v2 components into v1
