+++ packages/frontend/src/components/common/ShareSpaceModal/ShareSpaceModalContent.module.css
