+++ packages/frontend/src/components/common/ShareSpaceModal/ShareSpaceModalShared.module.css
